consolidate converter modules, fix dim name nnodes -> nodes
