+++ flopy4/mf6/converter.py
@@ -3,12 +3,19 @@ from datetime import datetime
 from pathlib import Path
 from typing import Any
 
+import numpy as np
+import sparse
 import xarray as xr
 import xattree
 from attrs import define
 from cattrs import Converter
+from numpy.typing import NDArray
+from xattree import get_xatspec
 
+from flopy4.adapters import get_nn
 from flopy4.mf6.component import Component
+from flopy4.mf6.config import SPARSE_THRESHOLD
+from flopy4.mf6.constants import FILL_DNODATA
 from flopy4.mf6.context import Context
 from flopy4.mf6.exchange import Exchange
 from flopy4.mf6.model import Model
@@ -148,7 +155,7 @@ def unstructure_component(value: Component) -> dict[str, Any]:
                 blocks[block_name][field_name] = tuple(field_value.values.tolist())
             elif isinstance(field_value, xr.DataArray) and "nper" in field_value.dims:
                 has_spatial_dims = any(
-                    dim in field_value.dims for dim in ["nlay", "nrow", "ncol", "nnodes"]
+                    dim in field_value.dims for dim in ["nlay", "nrow", "ncol", "nodes"]
                 )
                 if has_spatial_dims:
                     # terrible hack to convert flat nodes dimension to 3d structured dims.
@@ -156,7 +163,7 @@ def unstructure_component(value: Component) -> dict[str, Any]:
                     # should then have access to all dimensions needed.
                     dims_ = set(field_value.dims).copy()
                     dims_.remove("nper")
-                    if dims_ == {"nnodes"}:
+                    if dims_ == {"nodes"}:
                         parent = value.parent  # type: ignore
                         field_value = xr.DataArray(
                             field_value.data.reshape(
@@ -228,3 +235,73 @@ def _make_converter() -> Converter:
 
 
 COMPONENT_CONVERTER = _make_converter()
+
+
+def dict_to_array(value, self_, field) -> NDArray:
+    """
+    Convert a sparse dictionary representation of an array to a
+    dense numpy array or a sparse COO array.
+    """
+
+    if not isinstance(value, dict):
+        # if not a dict, assume it's a numpy array
+        # and let xarray deal with it if it isn't
+        return value
+
+    spec = get_xatspec(type(self_)).flat
+    field = spec[field.name]
+    if not field.dims:
+        raise ValueError(f"Field {field} missing dims")
+
+    # resolve dims
+    explicit_dims = self_.__dict__.get("dims", {})
+    inherited_dims = dict(self_.parent.data.dims) if self_.parent else {}
+    dims = inherited_dims | explicit_dims
+    shape = [dims.get(d, d) for d in field.dims]
+    unresolved = [d for d in shape if isinstance(d, str)]
+    if any(unresolved):
+        raise ValueError(f"Couldn't resolve dims: {unresolved}")
+
+    if np.prod(shape) > SPARSE_THRESHOLD:
+        a: dict[tuple[Any, ...], Any] = dict()
+
+        def set_(arr, val, *ind):
+            arr[tuple(ind)] = val
+
+        def final(arr):
+            coords = np.array(list(map(list, zip(*arr.keys()))))
+            return sparse.COO(
+                coords,
+                list(arr.values()),
+                shape=shape,
+                fill_value=field.default or FILL_DNODATA,
+            )
+    else:
+        a = np.full(shape, FILL_DNODATA, dtype=field.dtype)  # type: ignore
+
+        def set_(arr, val, *ind):
+            arr[ind] = val
+
+        def final(arr):
+            arr[arr == FILL_DNODATA] = field.default or FILL_DNODATA
+            return arr
+
+    if "nper" in dims:
+        for kper, period in value.items():
+            if kper == "*":
+                kper = 0
+            match len(shape):
+                case 1:
+                    set_(a, period, kper)
+                case _:
+                    for cellid, v in period.items():
+                        nn = get_nn(cellid, **dims)
+                        set_(a, v, kper, nn)
+            if kper == "*":
+                break
+    else:
+        for cellid, v in value.items():
+            nn = get_nn(cellid, **dims)
+            set_(a, v, nn)
+
+    return final(a)
