+++ flopy4/mf6/gwf/drn.py
@@ -7,7 +7,7 @@ from numpy.typing import NDArray
 from xattree import xattree
 
 from flopy4.mf6.component import update_maxbound
-from flopy4.mf6.converters import dict_to_array
+from flopy4.mf6.converter import dict_to_array
 from flopy4.mf6.package import Package
 from flopy4.mf6.spec import array, field
 
@@ -29,7 +29,7 @@ class Drn(Package):
     maxbound: Optional[int] = field(block="dimensions", default=None, init=False)
     elev: Optional[NDArray[np.float64]] = array(
         block="period",
-        dims=("nper", "nnodes"),
+        dims=("nper", "nodes"),
         default=None,
         converter=Converter(dict_to_array, takes_self=True, takes_field=True),
         reader="urword",
@@ -37,7 +37,7 @@ class Drn(Package):
     )
     cond: Optional[NDArray[np.float64]] = array(
         block="period",
-        dims=("nper", "nnodes"),
+        dims=("nper", "nodes"),
         default=None,
         converter=Converter(dict_to_array, takes_self=True, takes_field=True),
         reader="urword",
@@ -47,7 +47,7 @@ class Drn(Package):
         block="period",
         dims=(
             "nper",
-            "nnodes",
+            "nodes",
         ),
         default=None,
         converter=Converter(dict_to_array, takes_self=True, takes_field=True),
@@ -58,7 +58,7 @@ class Drn(Package):
         block="period",
         dims=(
             "nper",
-            "nnodes",
+            "nodes",
         ),
         default=None,
         converter=Converter(dict_to_array, takes_self=True, takes_field=True),
