+++ flopy4/mf6/gwf/rch.py
@@ -7,7 +7,7 @@ from numpy.typing import NDArray
 from xattree import xattree
 
 from flopy4.mf6.component import update_maxbound
-from flopy4.mf6.converters import dict_to_array
+from flopy4.mf6.converter import dict_to_array
 from flopy4.mf6.package import Package
 from flopy4.mf6.spec import array, field
 
@@ -29,7 +29,7 @@ class Rch(Package):
         block="period",
         dims=(
             "nper",
-            "nnodes",
+            "nodes",
         ),
         default=None,
         converter=Converter(dict_to_array, takes_self=True, takes_field=True),
@@ -40,7 +40,7 @@ class Rch(Package):
         block="period",
         dims=(
             "nper",
-            "nnodes",
+            "nodes",
         ),
         default=None,
         converter=Converter(dict_to_array, takes_self=True, takes_field=True),
@@ -51,7 +51,7 @@ class Rch(Package):
         block="period",
         dims=(
             "nper",
-            "nnodes",
+            "nodes",
         ),
         default=None,
         converter=Converter(dict_to_array, takes_self=True, takes_field=True),
